+++ packages/chaingraph-nodes/src/nodes/transforms/array/index.ts
@@ -6,16 +6,16 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
+export * from './array-concat.node'
+export * from './array-element.node'
 export * from './array-first.node'
 export * from './array-last.node'
-export * from './array-slice.node'
-export * from './array-concat.node'
+export * from './array-length.node'
+export * from './array-partition.node'
 export * from './array-pick.node'
 export * from './array-pluck.node'
-export * from './array-reverse.node'
-export * from './array-zip.node'
 export * from './array-reduce.node'
-export * from './array-partition.node'
-export * from './array-length.node'
+export * from './array-reverse.node'
+export * from './array-slice.node'
+export * from './array-zip.node'
 export * from './filter-array-lucene.node'
-export * from './array-element.node'