+++ packages/chaingraph-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @badaitech/chaingraph-types
 
+## 0.6.34
+
+### Patch Changes
+
+- 50796a3: feat(nodes): add Transforms category with 16 array/type/object nodes
+
 ## 0.6.33
 
 ### Patch Changes